test: Move full-environment test from liveslots to SwingSet
